+++ src/app/core/services/calendar-events.service.ts
@@ -30,20 +30,18 @@ export class CalendarEventsService {
 
   getCalendarEventsByRange(dateMinStr: string, dateMaxStr: string, entityId?: string): Observable<CalendarEvent[]> {
     const events$ = this.eventsSrv.getAllEvents(true, false, null, entityId);
-    const appointments$ = this.appointmentSrv.getAppointmentsByRange(
-      dateMinStr, dateMaxStr
-    );
+    const appointments$ = this.appointmentSrv.getAppointmentsByRange(dateMinStr, dateMaxStr, true);
 
     return combineLatest([
       appointments$,
       events$
     ])
       .pipe(
-        tap(([appointments, events ]) => {
-          // console.log(`Nº appointments: ${appointments.length}`);
-          // console.log(`Nº events: ${events.length}`);
-          // appointments.forEach(item => console.warn(item.id));
-        }),
+        // tap(([appointments, events ]) => {
+        //   console.log(`Nº appointments: ${appointments.length}`);
+        //   console.log(`Nº events: ${events.length}`);
+        //   appointments.forEach(item => console.warn(item.id));
+        // }),
         map(([appointments, events ]) => appointments
 
           .map(appointment =>