+++ src/app/shared/components/stories-panel/stories-panel.component.ts
@@ -10,6 +10,7 @@ import { IBase, BaseType, Base } from '@models/base';
 import { EventService } from '@services/events.service';
 import { LinksService } from '@services/links.services';
 import { NoticeService } from '@services/notices.service';
+import { StoriesService } from '@services/stories.service';
 import { LogService } from '@services/log.service';
 
 @Component({
@@ -23,11 +24,13 @@ export class StoriesPanelComponent implements OnInit {
   public notices$: Observable<INotice[]>;
   public events$: Observable<IEvent[]>;
   public REAL_STORIES: IBase[];
+  public stories$: Observable<IBase[]>;
   private listOfObservers: Array<Subscription> = [];
 
   constructor(
     private router: Router,
     private logSrv: LogService,
+    private storiesSrv: StoriesService,
     private noticesSrv: NoticeService,
     private linksSrv: LinksService,
     private eventsSrv: EventService,
@@ -35,52 +38,54 @@ export class StoriesPanelComponent implements OnInit {
 
   ngOnInit(): void {
 
-    this.links$ = this.linksSrv.getAllLinks(true, true, 2);
-    this.notices$ = this.noticesSrv.getAllNotices(true, true, 2);
-    this.events$ = this.eventsSrv.getAllEvents(true, true, 2, null);
+    this.stories$ = this.storiesSrv.getStories();
 
-    const subs1$ = combineLatest([
-      this.links$,
-      this.notices$,
-      this.events$,
-    ])
-    .subscribe(([links, notices, events]) => {
+    // this.links$ = this.linksSrv.getAllLinks(true, true, 2);
+    // this.notices$ = this.noticesSrv.getAllNotices(true, true, 2);
+    // this.events$ = this.eventsSrv.getAllEvents(true, true, 2, null);
 
-      this.REAL_STORIES = [];
+    // const subs1$ = combineLatest([
+    //   this.links$,
+    //   this.notices$,
+    //   this.events$,
+    // ])
+    // .subscribe(([links, notices, events]) => {
 
-      links.forEach(link => {
-        this.REAL_STORIES.push({
-          ...link,
-          baseType: BaseType.LINK,
-        });
-        // this.logSrv.info(`news item story! ${JSON.stringify(newsItem)}`);
-      });
+    //   this.REAL_STORIES = [];
 
-      notices.forEach(notice => {
-        this.REAL_STORIES.push({
-          ...notice,
-          baseType: BaseType.NOTICE,
-        });
-        // this.logSrv.info(`notice story! ${JSON.stringify(notice)}`);
-      });
+    //   // links.forEach(link => {
+    //   //   this.REAL_STORIES.push({
+    //   //     ...link,
+    //   //     baseType: BaseType.LINK,
+    //   //   });
+    //   //   // this.logSrv.info(`news item story! ${JSON.stringify(newsItem)}`);
+    //   // });
 
-      events.forEach(event => {
-        this.REAL_STORIES.push({
-          ...event,
-          baseType: BaseType.EVENT,
-        });
-        // this.logSrv.info(`event story! ${JSON.stringify(event)}`);
-      });
+    //   // notices.forEach(notice => {
+    //   //   this.REAL_STORIES.push({
+    //   //     ...notice,
+    //   //     baseType: BaseType.NOTICE,
+    //   //   });
+    //   //   // this.logSrv.info(`notice story! ${JSON.stringify(notice)}`);
+    //   // });
 
-      this.REAL_STORIES = this.REAL_STORIES.sort((item1: IBase, item2: IBase) => {
-        if (item1.timestamp > item2.timestamp) { return -1; }
-        if (item1.timestamp < item2.timestamp) { return 1; }
-        return 0;
-      });
+    //   events.forEach(event => {
+    //     this.REAL_STORIES.push({
+    //       ...event,
+    //       baseType: BaseType.EVENT,
+    //     });
+    //     // this.logSrv.info(`event story! ${JSON.stringify(event)}`);
+    //   });
 
-    });
+    //   this.REAL_STORIES = this.REAL_STORIES.sort((item1: IBase, item2: IBase) => {
+    //     if (item1.timestamp > item2.timestamp) { return -1; }
+    //     if (item1.timestamp < item2.timestamp) { return 1; }
+    //     return 0;
+    //   });
 
-    this.listOfObservers.push( subs1$ );
+    // });
+
+    // this.listOfObservers.push( subs1$ );
   }
 
   gotoItem(story: IBase): void {